+++ packages/widget/src/index.ts
@@ -1,7 +1,7 @@
 import type { Collection } from './collection-schema'
 import type { Widget } from './widget-schema'
 import path from 'node:path'
-import { readAndValidateJSONFile } from '@readme-widget-hub/utils'
+import { readAndValidateJSONFile, readAndValidateYamlFile } from '@readme-widget-hub/utils'
 import fs from 'fs-extra'
 import { collectionRootSchema } from './collection-schema'
 import { widgetRootSchema } from './widget-schema'
@@ -10,7 +10,8 @@ export * from './collection-schema'
 export * from './widget-schema'
 
 export const defaultIndex: number = 9999
-export const collectionFileName: string = 'collection.json'
+export const collectionFileName: string = 'collection'
+export const widgetOrCollectionExts: string[] = ['.json', '.yaml', '.yml']
 
 export interface WidgetNode extends Widget {
   type: 'widget'
@@ -26,13 +27,27 @@ export interface CollectionNode extends Collection {
 
 export type WidgetTree = WidgetNode | CollectionNode
 
+function isCollection(file: fs.Dirent): boolean {
+  const { name, ext } = path.parse(file.name)
+  return name === collectionFileName && widgetOrCollectionExts.includes(ext)
+}
+
+function tryGetCollection(dir: string): string | undefined {
+  for (const ext of widgetOrCollectionExts) {
+    const filePath = path.join(dir, `${collectionFileName}${ext}`)
+    if (fs.pathExistsSync(filePath)) {
+      return filePath
+    }
+  }
+}
+
 export function readWidgetTrees(dir: string, level: number = 1): WidgetTree[] {
   let nodes: WidgetTree[] = []
 
   for (const file of fs.readdirSync(dir, { withFileTypes: true })) {
     const filePath = path.join(dir, file.name)
 
-    if (file.name === collectionFileName) {
+    if (isCollection(file)) {
       continue
     }
 
@@ -53,8 +68,8 @@ export function readWidgetTrees(dir: string, level: number = 1): WidgetTree[] {
     .sort((a, b) => a.type === b.type ? 0 : a.type === 'collection' ? -1 : 1)
     .sort((a, b) => (a.index || defaultIndex) - (b.index || defaultIndex))
 
-  const collectionFilePath = path.join(dir, collectionFileName)
-  if (fs.pathExistsSync(collectionFilePath)) {
+  const collectionFilePath = tryGetCollection(dir)
+  if (collectionFilePath) {
     const collection = readCollection(collectionFilePath)
     if (collection === undefined || collection.enabled === false) {
       return []
@@ -77,9 +92,15 @@ export function flatWidgetTrees(tree: WidgetTree[]): WidgetTree[] {
 }
 
 export function readWidget(filePath: string): Widget {
+  if (filePath.endsWith('.yaml') || filePath.endsWith('.yml')) {
+    return readAndValidateYamlFile<Widget>(filePath, widgetRootSchema)
+  }
   return readAndValidateJSONFile<Widget>(filePath, widgetRootSchema)
 }
 
 export function readCollection(filePath: string): Collection {
+  if (filePath.endsWith('.yaml') || filePath.endsWith('.yml')) {
+    return readAndValidateYamlFile<Widget>(filePath, collectionRootSchema)
+  }
   return readAndValidateJSONFile<Widget>(filePath, collectionRootSchema)
 }