+++ packages/manager/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @readme-widget-hub/manager
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @readme-widget-hub/widget@2.2.0
+  - @readme-widget-hub/utils@2.1.0
+  - @readme-widget-hub/meta@2.0.2
+
 ## 2.0.2
 
 ### Patch Changes